avoid warning when compiling with -D_C2X_DEBUG

+++ cecilia2xkaapi/c2x.h
@@ -35,7 +35,7 @@
 #endif
 
 
-typedef long c2x_workqueue_index_t;
+typedef int c2x_workqueue_index_t;
 
 typedef struct {
   volatile c2x_workqueue_index_t beg __attribute__((aligned(64)));
